feat: some improvements
